monotonic clock: introduce `duration` type, split `subscribe` (#7358)
* monotonic clock: introduce `duration` type, split `subscribe`

We are introducing a `duration` type because it has a distinct meaning
from `instant`: an `instant` can only be compared to other `instant`s
from the exact same `monotonic-clock`, whereas a `duration` represents
a duration of time which can be compared to any other duration of time.

The `duration` type is motivated, in part, by a desire to reuse it to
specify durations such as timeouts in other WASI proposals.

Instead of taking a boolean specifying whether the u64 is an absolute or
relative time, `subscribe-instant` takes an `instant` type and
`subscribe-duration` takes a `duration` type.

* wasmtime-wasi: implement changes to monotonic clock interface

* adapter: fixes for subscribe duration and instant

* factor subscribe_to_duration out into a helper function.

* sync wits into wasi-http

* component adapter: fix for interpreting flag for monotonic instant vs duration

* subscribe_instant impl: fix logic for when time is in the past

caught by subtract overflow panic, thankfully

* test-programs: compile against new subscribe instant, duration methods