Attempt to automatically configure release notes (#7405) (#7424)
* Attempt to automatically configure release notes

This commit is an attempt to tackle #7068 by configuring the release
notes in Github Releases with the handwritten release notes from
`RELEASES.md`. The basic idea here is to split the markdown file on
`-----` delimiters and then find the one which matches the version being
released. Once one is found the `body` field of the API call to create
the release is configured.

* Update .github/actions/github-release/main.js



---------

Co-authored-by: Andrew Brown <andrew.brown@intel.com>