add `reset_adapter_state` export to adapter
This is useful as the last step of component pre-initialization with
e.g. [component-init](https://github.com/dicej/component-init), in which case we
want the adapter to forget about any open handles it has (and force it to
re-request the stdio handles next time they're needed) since it will be talking
to a brand new host at runtime.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>