c-api: support IWYU pragmas (#7473)
This allows clangd to not import the headers in the wasmtime directory,
as not all of them include their dependencies correctly. This means that
clangd can import a header directly and it breaks the build, as the
order `wasmtime.h` #includes is important.

See: https://github.com/include-what-you-use/include-what-you-use/blob/master/docs/IWYUPragmas.md

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>