zkasm: implement lowering for bnot
This is binary not (~x in C or !x in Rust.) WASM does not produce this
specific instruction and lowers it to `xor` usually, but in zkasm it is
better for us to use subtraction instead – as it is faster.

I believe this should produce equivalent results for all inputs.