winch: Multi-Value Part 1  (#7535)
* winch: Introduce `ABIParams` and `ABIResults`

This commit prepares Winch to support WebAssembly Multi-Value.

The most notorious piece of this change is the introduction of the
`ABIParams` and `ABIResults` structs which are type wrappers around the
concept of an `ABIOperand`, which is the underlying main representation
of a param or result.

This change also consolidates how the size for WebAssembly types is
derived by introducing `ABI::sizeof`, as well as introducing
`ABI::stack_slot_size` to concretely indicate the stack slot size in
bytes for stack params, which is ABI dependent.

* winch: Add the necessary ABI building blocks for multi-value

This change adds the necessary changes at the ABI level in order to
handle multi-value.

The most notable modifications in this change are:

* Modifying Winch's  default ABI to reverse the order of results,
  ensuring that results that go in the stack should always come first;
  this makes it easier to respect the following two stack invariants:

  * Spilled memory values always precede register values
  * Spilled values are stored from oldest to newest, matching their
    respective locations on the machine stack.

* Modify all calling conventions supported by Winch so that only one result, the first one is stored in
  registers. This differs from their vanilla counterparts in that these
  ABIs can handle multiple results in registers. Given that Winch is not
  a generic code generator, keeping the ABI close to what Wasmtime
  expects makes it easier to pass multiple results at trampolines.

* Add more multi-value tests

This commit adds more tests for multi-value and improves documentation.

prtest:full

* Address review feedback