Add zkevm-proverjs as a submodule
This simplifies the development setup both locally and on CI, as we
don't need to clone zkevm-proverjs as an explicit step.

Given that we don't plan to directly merge back into wasmtime upstream,
I think it's fine to deviate in more major ways from the existing test
infrastructure if this gives us productivity gains.