Wasi test suite cleanups (#7553)
* rename preview1_remove_directory_trailing_slashes to preview1_remove_directory

and get rid its testing of the inconsistient behavior of removing a
directory using a trailing slash.

Then, unmark it as should_panic everywhere, so we have test coverage of
the behaviors that are consistient

prtest:full

* delete preview1_path_rename_file_trailing_slashes

this only tests behavior which is not consistient across platforms. the
rest of the behavior of path_rename is covered well by the path_rename
test and path_rename_dir_trailing_slashes

* debugging: see what CI tells me the preview1_interesting_paths chokes on in windows

* more debugging information

* interesting paths: windows allows a trailing nul apparently

* preview1_remove_directory: every os says notdir, apparently!

* no more should_panic tests!