ci: log CPU details when testing (#7574)
* ci: log CPU details when testing

When testing, there are certain CPU-dependent features that influence
Cranelift's codegen (e.g., availability of AVX512 instructions). This
additional CI step logs the current CPU information to aid in
troubleshooting, such as the MPK-related troubleshooting over in #7445.
Also, if we let this run in CI for a while, we may be able to run
queries on the logs to determine how often jobs run on servers with
certain features enabled.

prtest:full

* Add Windows variant of 'lscpu'

* Add MacOS variant of 'lscpu'