winch: Ensure stack pointer for br_table
Follow up to:

https://github.com/bytecodealliance/wasmtime/pull/7547

In which I overlooked this change and the fuzzer found an issue with the
following program:

```wat
(module
  (func (export "") (result i32)
    block (result i32)
       i32.const 0
    end
    i32.const 0
    i32.const 0
    br_table 0
  )
)
```

This commit ensures that the stack pointer is correctly positioned when
emitting br_table.

We can't know for sure which branch will be taken, but since all
branches must share the same type information, we can be certain that
the expectations regarding the stack pointer are the same and thus can
we use the default target in order to ensure the correct placement.