winch: Rework `br_table` jumps
This commit reworks the `br_table` logic so that it correctly handles
all the jumps involved to each of the targets.

Even though it is safe to use the default branch for type information,
it is not safe to use it to derive the base stack pointer and base value
stack length. This change ensures that each target offset is taken into
account to balance the value stack prior to each jump.