[WASI] Dynamic `SocketAddr` check for outbound socket traffic (#7662)
* Add API to WasiCtx for adding dynamic SocketAddr checks

The user can now specify a closure which gets run any time a connection
to an external address is attempted. If the closure returns true, the
address is allowed, but if it returns false, the underlying pool is then
checked. In otherwords, false does not imply denied, but rather, check
the underlyin pool.

Signed-off-by: Ryan Levick <ryan.levick@fermyon.com>

* Get rid of Pool in favor of just a closure

Signed-off-by: Ryan Levick <ryan.levick@fermyon.com>

* Add a SocketAddrUse arg to the SocketAddrCheck

Signed-off-by: Ryan Levick <ryan.levick@fermyon.com>

---------

Signed-off-by: Ryan Levick <ryan.levick@fermyon.com>