This follows the design outlined in wasmtime#7681 for IpNameLookup.
A new trait `IpNameLookup` is introduced that factors out the conversion
from a `String` to a list of IP addresses. Another trait,
`WasiIpNameLookupView` is used to configure which concrete type that
implements `IpNameLookup` will be used in the actual IpNameLookup host
implementation. Lastly. the `SystemIpNameLookup` is provided as a
default implementation.

Signed-off-by: Ryan Levick <ryan.levick@fermyon.com>