winch: Multi-Value Part 2: Blocks (#7707)
* winch: Multi-Value Part 2: Blocks

This commit adds support for the Multi-Value proposal for blocks.

In general, this change, introduces multiple building blocks to enable
supporting arbitrary params and results in blocks:

* `BlockType`: Introduce a block type, to categorize the type of each
  block, this makes it easier to categorize blocks per type and also
  makes it possible to defer the calculation of the `ABIResults` until
  they are actually needed rather than calculating everyghing upfront
  even though they might not be needed (when in an unreachable state).
* Push/pop operations are now frame aware. Given that each
  `ControlStackFrame` contains all the information needed regarding
  params and results, this change moves the the implementation of the
  push and pop operations to the `ControlStackFrame` struct.
* `StackState`: this struct holds the entry and exit invariants of each
  block; these invariants are pre-computed when entering the block and
  used throughout the code generation, to handle params, results and
  assert the respective invariants.

In terms of the mechanics of the implementation: when entering each
block, if there are results on the stack, the expected stack pointer
offsets will be calculated via the `StackState`, and the `target_offset`
will be used to create the block's `RetArea`. Note that when entering
the block and calculating the `StackState` no space is actually reserved
for the results, any space increase in the stack is deffered until the
results are popped from the value stack via
`ControlStackFrame::pop_abi_results`.

The trickiest bit of the implementation is handling constant values that
need to be placed on the right location on the machine stack. Given that
constants are generally not spilled, this means that in order to keep
the machine and value stack in sync (spilled-values-wise), values must
be shuffled to ensure that constants are placed in the expected location results wise.
See the comment in `ControlStackFrame::adjust_stack_results` for more
details.

* Review fixes